[DF-4355] Fixed opening of png files
Former-commit-id: af80d7431249a9507cbbaa8febb7e53530eedaeb

+++ Sources/Internal/Render/Image/ImageConvert.h
@@ -138,8 +138,7 @@ struct ConvertA8toRGBA8888
 {
 	inline void operator()(const uint8 * input, uint32 *output)
 	{
-		uint8 pixel = *input;
-
+		uint32 pixel = *input;
 		*output = ((pixel) << 24) | (pixel << 16) | (pixel << 8) | pixel;
 	}
 };
